Moved commands to cmd module
