+++ example_radar.yml
@@ -1,0 +1,68 @@
+types:
+  - datastore
+  - data_management
+  - infrastructure
+  - language
+  - frameworks
+
+statuses:
+  - hold
+  - assess
+  - trial
+  - adopt
+
+technologies:
+  - name: Python
+    type: language
+    status: adopt
+    description: >
+      Python is a high-level, general-purpose programming language. Its design 
+      philosophy emphasizes code readability with the use of significant 
+      indentation. Python is dynamically-typed and garbage-collected. It 
+      supports multiple programming paradigms, including structured, 
+      object-oriented and functional programming.
+    website: https://python.org
+  - name: Golang
+    type: language
+    status: assess
+    description: >
+      Go is a statically typed, compiled programming language designed at Google
+       by Robert Griesemer, Rob Pike, and Ken Thompson. It is syntactically 
+       similar to C, but with memory safety, garbage collection, structural 
+       typing, and CSP-style concurrency.
+    website: https://go.dev
+  - name: Django
+    type: framework
+    status: hold
+    description: >
+      Django is a free and open-source, Python-based web framework that follows 
+      the model–template–views architectural pattern. It is maintained by the 
+      Django Software Foundation, an independent organization established in the
+       US as a 501 non-profit.
+    website: https://www.djangoproject.com
+  - name: FastAPI
+    type: framework
+    status: adopt
+    description: >
+      FastAPI is a Web framework for developing RESTful APIs in Python. FastAPI 
+      is based on Pydantic and type hints to validate, serialize, and 
+      deserialize data, and automatically auto-generate OpenAPI documents.
+    website: https://fastapi.tiangolo.com
+  - name: PostgreSQL
+    type: datastore
+    status: adopt
+    description: >
+      PostgreSQL, also known as Postgres, is a free and open-source relational 
+      database management system emphasizing extensibility and SQL compliance. 
+      It was originally named POSTGRES, referring to its origins as a successor 
+      to the Ingres database developed at the University of California, 
+      Berkeley.
+    website: https://www.postgresql.org
+  - name: Airflow
+    type: data_management
+    status: trial
+    description: >
+      Apache Airflow is an open-source workflow management platform for data 
+      engineering pipelines. It started at Airbnb in October 2014 as a solution 
+      to manage the company's increasingly complex workflows.
+    website: https://airflow.apache.org
